extract database setup to a parent class and just inhereit
for resusablility
